Fix [#3] extension 폴더링
