+++ 32nd_SOPT_First_Seminar/32nd_SOPT_First_Seminar/Common/Extension/UIButton+.swift
@@ -1,0 +1,20 @@
+//
+//  UIButton+.swift
+//  32nd_SOPT_Seminar
+//
+//  Created by 정채은 on 2023/04/11.
+//
+
+import UIKit.UIButton
+
+extension UIButton {
+    func setUnderline() {
+        guard let title = title(for: .normal) else { return }
+        let attributedString = NSMutableAttributedString(string: title)
+        attributedString.addAttribute(.underlineStyle,
+                                      value: NSUnderlineStyle.single.rawValue,
+                                      range: NSRange(location: 0, length: title.count)
+        )
+        setAttributedTitle(attributedString, for: .normal)
+    }
+}
