+++ 32nd_SOPT_First_Seminar/32nd_SOPT_First_Seminar/Common/Extension/UITextField+.swift
@@ -1,0 +1,23 @@
+//
+//  UITextField+.swift
+//  32nd_SOPT_Seminar
+//
+//  Created by 정채은 on 2023/04/11.
+//
+
+import UIKit.UITextField
+
+extension UITextField {
+    func setPlaceholder(color: UIColor) {
+        guard let string = self.placeholder else {
+            return
+        }
+        attributedPlaceholder = NSAttributedString(string: string, attributes: [.foregroundColor: color])
+    }
+    
+    func addLeftPadding() {
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: 22, height: self.frame.height))
+        self.leftView = paddingView
+        self.leftViewMode = ViewMode.always
+    }
+}
